add margin calculator from utility
